+++ packages/flutter_timeline/lib/src/flutter_timeline_gorouter_userstory.dart
@@ -50,7 +50,7 @@ List<GoRoute> getTimelineStoryRoutes({
                   .optionsBuilder(context)
                   .theme
                   .postCreationFloatingActionButtonColor ??
-              const Color(0xff71C6D1),
+              Theme.of(context).primaryColor,
           onPressed: () async => context.push(
             TimelineUserStoryRoutes.timelineCategorySelection,
           ),
@@ -75,8 +75,8 @@ List<GoRoute> getTimelineStoryRoutes({
                         .optionsBuilder(context)
                         .translations
                         .timeLineScreenTitle!,
-                    style: const TextStyle(
-                      color: Color(0xff71C6D1),
+                    style: TextStyle(
+                      color: Theme.of(context).primaryColor,
                       fontSize: 24,
                       fontWeight: FontWeight.w800,
                     ),
@@ -121,8 +121,8 @@ List<GoRoute> getTimelineStoryRoutes({
                   backgroundColor: const Color(0xff212121),
                   title: Text(
                     config.optionsBuilder(context).translations.postCreation!,
-                    style: const TextStyle(
-                      color: Color(0xff71C6D1),
+                    style: TextStyle(
+                      color: Theme.of(context).primaryColor,
                       fontSize: 24,
                       fontWeight: FontWeight.w800,
                     ),
@@ -178,8 +178,8 @@ List<GoRoute> getTimelineStoryRoutes({
                   backgroundColor: const Color(0xff212121),
                   title: Text(
                     post.category ?? 'Category',
-                    style: const TextStyle(
-                      color: Color(0xff71C6D1),
+                    style: TextStyle(
+                      color: Theme.of(context).primaryColor,
                       fontSize: 24,
                       fontWeight: FontWeight.w800,
                     ),
@@ -237,8 +237,8 @@ List<GoRoute> getTimelineStoryRoutes({
                   leading: backButton,
                   title: Text(
                     config.optionsBuilder(context).translations.postCreation!,
-                    style: const TextStyle(
-                      color: Color(0xff71C6D1),
+                    style: TextStyle(
+                      color: Theme.of(context).primaryColor,
                       fontSize: 24,
                       fontWeight: FontWeight.w800,
                     ),
@@ -285,8 +285,8 @@ List<GoRoute> getTimelineStoryRoutes({
                   backgroundColor: const Color(0xff212121),
                   title: Text(
                     config.optionsBuilder(context).translations.postOverview!,
-                    style: const TextStyle(
-                      color: Color(0xff71C6D1),
+                    style: TextStyle(
+                      color: Theme.of(context).primaryColor,
                       fontSize: 24,
                       fontWeight: FontWeight.w800,
                     ),