+++ packages/flutter_timeline_view/lib/src/config/timeline_theme.dart
@@ -16,6 +16,7 @@ class TimelineTheme {
     this.moreIcon,
     this.deleteIcon,
     this.categorySelectionBorderColor,
+    this.categorySelectionButtonBackgroundColor,
     this.postCreationFloatingActionButtonColor,
     this.textStyles = const TimelineTextStyles(),
   });
@@ -45,6 +46,10 @@ class TimelineTheme {
 
   /// The color of the border around the category in the selection screen
   final Color? categorySelectionBorderColor;
+
+  /// The color of the background of the category selection button in the
+  /// selection screen
+  final Color? categorySelectionButtonBackgroundColor;
 
   /// The color of the floating action button on the overview screen
   final Color? postCreationFloatingActionButtonColor;