+++ packages/flutter_timeline_view/lib/src/widgets/category_selector.dart
@@ -37,7 +37,7 @@ class _CategorySelectorState extends State<CategorySelector> {
           SizedBox(
             width: widget.options.categoriesOptions
                     .categorySelectorHorizontalPadding ??
-                max(widget.options.padding.horizontal - 20, 0),
+                max(widget.options.padding.left - 20, 0),
           ),
           for (var category in categories) ...[
             widget.options.categoriesOptions.categoryButtonBuilder?.call(
@@ -61,7 +61,7 @@ class _CategorySelectorState extends State<CategorySelector> {
           SizedBox(
             width: widget.options.categoriesOptions
                     .categorySelectorHorizontalPadding ??
-                max(widget.options.padding.horizontal - 4, 0),
+                max(widget.options.padding.right - 4, 0),
           ),
         ],
       ),