+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add a post and a category to the postViewOpenPageBuilder function
 - Add a refresh functionality to the timeline with a pull to refresh callback to allow additional functionality when refreshing the timeline
 - Use the adaptive variants of the material elements in the timeline
+- Change the default blue color to the primary color of the Theme.of(context) in the timeline
 
 ## 3.0.1